feat: update path parser to handle all path types
